make intermediary component for zendesk wrapper
